Fix build scripts (#5)
The `prepublishOnly` script was broken due to poorly composed build scripts. In particular, `build:clean` tried to remove the same file twice, causing an error. While fixing that issue, I decided to refactor the build scripts.

This PR refactors the build scripts to only minimally use Bash, and to not break during releases. In addition, the raw SLIP-44 data is now committed to this repository, both for clerical purposes and in order to stop fetching it in CI. Finally, a clean working directory check is added to CI after building from the committed markdown.